+++ public/index.html
@@ -29,7 +29,7 @@
 
   <!-- //external files created by Mitali Patel -->
   <script type="module" src="js/pointcloud.js" ></script>
-  <script type="module" src="js/callingFunc.js"></script>
+  <script type="module" src="js/loading.js"></script>
   <!-- <script type="module" src="js/IntegrationCode.js" ></script> -->
   <script type="module" src="api-web-devs/widgets/loader.js"></script>
   <script type="module" src="api-web-devs/parsers/standardized.js"></script>
