+++ public/js/loading.js
